Refactor: move geometry logic into the base Row from GeoJsonRow
